requirements.txt
SQLAlchemy
joblib
numpy
pandas
plotly
psycopg2-binary
python-dotenv
scikit-learn
streamlit
tensorflow
